checkrt: Sort symbols
Some systems have out-of-order symbols in libstdc++, which fools the old
logic into thinking the installed libstdc++ library is older than the
included version. This causes problems with host libraries that need a
newer libstdc++ library.

Sort the symbols found in the library and use the last one.

checkrt: Fix symbol sorting